requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0
fiona>=1.9.0

# Data processing
pyyaml>=6.0
h5py>=3.8.0
netCDF4>=1.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
pillow>=9.5.0

# Machine Learning utilities
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
keras-tuner>=1.4.0

# Streamlit app
streamlit>=1.25.0
streamlit-folium>=0.13.0
folium>=0.14.0

# Progress bars and logging
tqdm>=4.65.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0
